More changes in preparation for emulating legacy mach_init servers as launchd jobs:
mach_init's data structures used to be arranged into three global lists:

All bootstraps.
All servers.
All services.

Each bootstrap had a pointer to the "parent" bootstrap.
Each service had a pointer to the bootstrap it was in.
Each server had a pointer back to the bootstrap it is in.

A given server's services were identified by a pointer in the service structure
back to the server structure.

This old design led to lots of O(n^2) and O(n^3) behavior when the relationship
between data structures needed to be validated or walked.

In this commit, we change the data structures to map to the algorithmic needs
of mach_init. Everything is O(n) now.

The only global we have now is a pointer to the root bootstrap.

Each bootstrap still has a pointer to the parent.
Each bootstrap now has three list hanging off of them.
    One for sub-bootstraps.
    One for servers in that bootstrap scope.
    One for serverless services in that bootstrap scope.

Each server still has a pointer to the bootstrap it is in.
Each server now has a list of services it vends.

Each service still has a pointer to the bootstrap it is in.
Each service still has a pointer to the server it belongs to, if any.


git-svn-id: http://svn.macosforge.org/repository/launchd/trunk@20476 e92bca22-270c-0410-9cea-e3f1106b6a1c