launchd uniquely identifies jobs by a string label.
When jobs are submitted via the legacy mach_init bootstrap_create_server() API,
we must construct a label to fit into the launchd universe. I have been
informed that using the word "legacy" in these automagic labels would be
premature at this point in time, so let's change the constructed label to be
more politically correct for the time being.


git-svn-id: http://svn.macosforge.org/repository/launchd/trunk@20522 e92bca22-270c-0410-9cea-e3f1106b6a1c