job_import() had grown unwieldy. This breaks it into bite size pieces.
This change not only improves the readability and maintainability, but
it also improves the robustness of job schema validation and the performance
of parsing the incoming job description.

Essentially, we've gone from looking up / testing for the existence of keys,
and then validating the type, to a model where we iterate over the dictionary,
and demux based on the object type of the value to the right import function.
The respective import-by-value-type functions only need to compare the key
against a limited set that are known to support that value type.


git-svn-id: http://svn.macosforge.org/repository/launchd/trunk@20560 e92bca22-270c-0410-9cea-e3f1106b6a1c