Structural and semantic reorganization:
When possible, let's have the services created with bootstrap_register() associated with the job that created them rather than the bootstrap itself.

This, in theory, could break things if the job is removed and it didn't own the receive right that it registered in the namespace, but in practice, I have not seen any evidence for this boundary condition being used.


git-svn-id: http://svn.macosforge.org/repository/launchd/trunk@22872 e92bca22-270c-0410-9cea-e3f1106b6a1c