Structural and semantic reorganization:
This change is fairly large, but it gets things going in the right direction. About November/December of 2005, I noticed that Mach bootstraps and launchd jobs had a lot in common from a data structure perspective, so I merged them. Almost one year later, I'm thankful for the learning experience. Suffice it for me to say, they aren't the same.

With this change, we now create "anonymous" jobs on the fly when the PID is not found in launchd's object graph.

What this means, is we'll be able to easily undo the merger of the bootstrap and job data structures, but not go back to the mess I was trying to solve at the end of 2005.


git-svn-id: http://svn.macosforge.org/repository/launchd/trunk@22876 e92bca22-270c-0410-9cea-e3f1106b6a1c